requirements.txt
langchain
langgraph
langchain-community
langchain-text-splitters
langchain-groq

chromadb

python-dotenv

streamlit

tqdm

sentence-transformers

torchvision
